bot-win/test/pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = "test"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_remake", "nonebot_plugin_simplemusic", "nonebot_plugin_setu", "nonebot_plugin_answersbook"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
